Moved handlers back to source/sink classes.
