+++ chronicle/src/main/java/net/openhft/chronicle/network/TcpEventHandler.java
@@ -24,19 +24,21 @@ public class TcpEventHandler {
 
     private final SocketChannel sc;
     private final TcpHandler handler;
-    private final ByteBuffer inBB = ByteBuffer.allocateDirect(CAPACITY);
-    private final ByteBuffer outBB = ByteBuffer.allocateDirect(CAPACITY);
+    private final ByteBuffer inBB;
+    private final ByteBuffer outBB;
     private final Bytes inBBB;
     private final Bytes outBBB;
     private final SessionDetailsProvider sessionDetails;
 
-//    public TcpEventHandler(@NotNull SocketChannel sc, TcpHandler handler, SessionDetails sessionDetails, boolean unchecked) throws IOException {
-    public TcpEventHandler(@NotNull SocketChannel sc, TcpHandler handler, SessionDetailsProvider sessionDetails) throws IOException {
+    public TcpEventHandler(@NotNull SocketChannel sc, TcpHandler handler, SessionDetailsProvider sessionDetails, int sendCapacity, int receiveCapacity) throws IOException {
+        inBB = ByteBuffer.allocateDirect(receiveCapacity > 0 ? receiveCapacity : CAPACITY);
+        outBB = ByteBuffer.allocateDirect(sendCapacity > 0 ? sendCapacity : CAPACITY);
+
         this.sc = sc;
         sc.configureBlocking(false);
         sc.socket().setTcpNoDelay(true);
-        sc.socket().setReceiveBufferSize(CAPACITY);
-        sc.socket().setSendBufferSize(CAPACITY);
+        sc.socket().setReceiveBufferSize(inBB.capacity());
+        sc.socket().setSendBufferSize(outBB.capacity());
 
         this.handler = handler;
         this.sessionDetails = sessionDetails;
@@ -49,6 +51,11 @@ public class TcpEventHandler {
         outBBB = wrap(outBB.slice());
         // must be set after we take a slice();
         outBB.limit(0);
+    }
+
+//    public TcpEventHandler(@NotNull SocketChannel sc, TcpHandler handler, SessionDetails sessionDetails, boolean unchecked) throws IOException {
+    public TcpEventHandler(@NotNull SocketChannel sc, TcpHandler handler, SessionDetailsProvider sessionDetails) throws IOException {
+        this(sc, handler, sessionDetails, CAPACITY, CAPACITY);
     }
 
     public boolean action() throws InvalidEventHandlerException {
@@ -76,6 +83,7 @@ public class TcpEventHandler {
     boolean invokeHandler() throws IOException {
         long start = System.nanoTime();
         boolean busy = false;
+        long inBBBPos = inBBB.position();
         // inBBB.readLimit(inBB.position());
         inBBB.limit(inBB.position());
         // outBBB.writePosition(outBB.limit());
@@ -88,7 +96,7 @@ public class TcpEventHandler {
             // outBB.limit(Maths.toInt32(outBBB.writePosition()));
             outBB.limit(Maths.toInt(outBBB.position(), "Int %d out of range"));
             tryWrite();
-            busy = true;
+            busy |= true;
         }
 
         long compact = System.nanoTime();
@@ -107,7 +115,9 @@ public class TcpEventHandler {
             inBBB.position(0);
             // inBBB.readLimit(inBB.position());
             inBBB.limit(inBB.position());
-            busy = true;
+            busy |= true;
+        } else {
+            busy |= inBBBPos < inBBB.position(); // we have read something, so deemed busy.
         }
         if (LOG.get()) {
             long now = System.nanoTime();
