+++ chronicle/src/main/java/net/openhft/chronicle/tcp/SourceTcp.java
@@ -18,14 +18,12 @@
 package net.openhft.chronicle.tcp;
 
 import net.openhft.chronicle.*;
-import net.openhft.chronicle.network.InvalidEventHandlerException;
-import net.openhft.chronicle.network.SimpleSessionDetailsProvider;
-import net.openhft.chronicle.network.TcpEventHandler;
+import net.openhft.chronicle.network.*;
 import net.openhft.chronicle.tools.ResizableDirectByteBufferBytes;
 import net.openhft.lang.io.Bytes;
-import net.openhft.lang.io.DirectByteBufferBytes;
 import net.openhft.lang.model.constraints.NotNull;
 import net.openhft.lang.thread.LightPauser;
+import net.openhft.lang.thread.Pauser;
 import org.jetbrains.annotations.Nullable;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
@@ -135,12 +133,9 @@ public abstract class SourceTcp {
         protected ExcerptTailer tailer;
         protected ExcerptAppender appender;
         protected long lastHeartbeat;
-        private long lastUnPausedNS;
 
         protected final Bytes bytesOut;
         protected final ResizableDirectByteBufferBytes readBuffer;
-
-        private ResizableDirectByteBufferBytes withMappedBuffer;
 
         private SessionHandler(final @NotNull SocketChannel socketChannel, SourceTcpHandler sourceTcpHandler) throws IOException {
             this.socketChannel = socketChannel;
@@ -150,18 +145,15 @@ public abstract class SourceTcp {
             this.sourceTcpHandler.setMaxExcerptsPerMessage(builder.maxExcerptsPerMessage());
             this.sourceTcpHandler.setHeartbeatIntervalMillis(builder.heartbeatIntervalMillis());
             this.sessionDetails = new SimpleSessionDetailsProvider();
-            this.tcpEventHandler = new TcpEventHandler(socketChannel, pipeline(sourceTcpHandler), sessionDetails);
+            this.tcpEventHandler = new TcpEventHandler(socketChannel, pipeline(sourceTcpHandler), sessionDetails, builder.sendBufferSize(), builder.receiveBufferSize());
             this.tailer = null;
             this.appender = null;
             this.lastHeartbeat = 0;
-            this.lastUnPausedNS = 0;
 
             this.readBuffer = new ResizableDirectByteBufferBytes(16);
             this.readBuffer.clearThreadAssociation();
             this.bytesOut = wrap(createBuffer(builder.minBufferSize()));
             this.bytesOut.limit(0);
-
-            this.withMappedBuffer = new ResizableDirectByteBufferBytes(1024);
         }
 
         @Override
@@ -170,7 +162,7 @@ public abstract class SourceTcp {
         }
 
         @Override
-        public void unsubscribed() {
+        public void unSubscribed() {
             final SelectionKey selectionKey = sessionDetails.get(SelectionKey.class);
             selectionKey.interestOps(selectionKey.interestOps() & ~SelectionKey.OP_WRITE);
         }
@@ -301,51 +293,6 @@ public abstract class SourceTcp {
             }
         }
 
-        protected boolean hasRoomForExcerpt(Bytes bytes, Bytes tailer) {
-            return hasRoomFor(bytes, tailer.remaining() + ChronicleTcp.HEADER_SIZE);
-        }
-
-        protected boolean hasRoomFor(Bytes bytes, long size) {
-            return bytes.remaining() >= size;
-        }
-
-        protected void pauseReset() {
-            lastUnPausedNS = System.nanoTime();
-            pauser.reset();
-        }
-
-        protected void pause() {
-            if (lastUnPausedNS + ChronicleTcp.BUSY_WAIT_TIME_NS > System.nanoTime()) {
-                return;
-            }
-
-            pauser.pause();
-        }
-
-        protected void setLastHeartbeat() {
-            this.lastHeartbeat = System.currentTimeMillis() + builder.heartbeatIntervalMillis();
-        }
-
-        protected void setLastHeartbeat(long from) {
-            this.lastHeartbeat = from + builder.heartbeatIntervalMillis();
-        }
-
-        protected void sendSizeAndIndex(int size, long index) throws IOException {
-            bytesOut.clear();
-            bytesOut.writeInt(size);
-            bytesOut.writeLong(index);
-            // TODO
-//            connection.write(bytesOut.flip());
-
-            setLastHeartbeat();
-        }
-
-        protected DirectByteBufferBytes readUpTo(int size) throws IOException {
-            // TODO
-//            connection.read(readBuffer.resetToSize(size), size, -1);
-            return readBuffer;
-        }
-
         protected boolean onSelectionKey(final SelectionKey key) throws IOException {
             if (key != null) {
                 if (key.isReadable() || key.isWritable()) {
@@ -359,120 +306,255 @@ public abstract class SourceTcp {
                         sessionDetails.set(SelectionKey.class, null);
                     }
                 }
-/*
-                    if (!onRead(key)) {
-                        return false;
-                    }
-                } else if (key.isWritable()) {
-                    if (!onWrite(key)) {
-                        return false;
-                    }
-                }
-*/
             }
 
             return true;
         }
 
-        protected boolean onRead(final SelectionKey key) throws IOException {
-            try {
-                final long action = readUpTo(8).readLong();
-                switch ((int) action) {
-                    case (int) ChronicleTcp.ACTION_WITH_MAPPING:
-                        return onMapping(key, readUpTo(4).readInt());
-                    case (int) ChronicleTcp.ACTION_SUBSCRIBE:
-                        return onSubscribe(key, readUpTo(8).readLong());
-                    case (int) ChronicleTcp.ACTION_UNSUBSCRIBE:
-                        return onUnsubscribe(key, readUpTo(8).readLong());
-                    case (int) ChronicleTcp.ACTION_QUERY:
-                        return onQuery(key, readUpTo(8).readLong());
-                    case (int) ChronicleTcp.ACTION_SUBMIT:
-                        return onSubmit(key, readUpTo(8).readLong(), true);
-                    case (int) ChronicleTcp.ACTION_SUBMIT_NOACK:
-                        return onSubmit(key, readUpTo(8).readLong(), false);
-                    default:
-                        throw new IOException("Unknown action received (" + action + ")");
-                }
-            } catch (IOException e) {
-                key.selector().close();
-                throw e;
-            }
+    }
+
+    // *************************************************************************
+    // SessionHandler - implementations
+    // *************************************************************************
+
+    /**
+     * IndexedChronicle session handler
+     */
+    private class IndexedSessionHandler extends SessionHandler {
+
+        private IndexedSessionHandler(final @NotNull SocketChannel socketChannel) throws IOException {
+            super(socketChannel, SourceTcpHandler.indexed());
         }
 
-        protected boolean onWrite(final SelectionKey key) throws IOException {
-            final long now = System.currentTimeMillis();
-            final Object attachment = key.attachment();
+    }
 
-            if (running.get() && !write(attachment)) {
-                if (lastHeartbeat <= now) {
-                    sendSizeAndIndex(ChronicleTcp.IN_SYNC_LEN, ChronicleTcp.IDX_NONE);
-                }
-            }
+    /**
+     * VanillaChronicle session handler
+     */
+    private class VanillaSessionHandler extends SessionHandler {
 
-            return true;
+        private VanillaSessionHandler(final @NotNull SocketChannel socketChannel) throws IOException {
+            super(socketChannel, SourceTcpHandler.vanilla());
         }
 
-        protected boolean onMapping(final SelectionKey key, int size) throws IOException {
-            MappingProvider mappingProvider = (MappingProvider) key.attachment();
-            if (mappingProvider != null) {
-                MappingFunction mappingFunction = readUpTo(size).readObject(MappingFunction.class);
-                mappingProvider.withMapping(mappingFunction);
-            }
+    }
 
-            return true;
+    public static abstract class SourceTcpHandler implements TcpHandler {
+
+        protected Pauser pauser;
+
+        protected ExcerptTailer tailer;
+
+        protected ExcerptAppender appender;
+
+        protected SubscriptionListener subscriptionListener;
+
+        private boolean subscribed;
+
+        private long lastHeartbeat;
+
+        protected int maxExcerptsPerMessage;
+
+        private long heartbeatIntervalMillis;
+
+        protected long index;
+
+        protected ResizableDirectByteBufferBytes withMappedBuffer;
+
+        protected TcpHandlerState state;
+
+        protected Bytes content;
+
+        public SourceTcpHandler() {
+            this.withMappedBuffer = new ResizableDirectByteBufferBytes(1024);
         }
 
-        protected boolean onQuery(final SelectionKey key, long data) throws IOException {
-            if (tailer.index(data)) {
-                final long now = System.currentTimeMillis();
-                setLastHeartbeat(now);
+        public void setTailer(ExcerptTailer tailer) {
+            this.tailer = tailer;
+        }
 
-                while (true) {
-                    if (tailer.nextIndex()) {
-                        sendSizeAndIndex(ChronicleTcp.SYNC_IDX_LEN, tailer.index());
-                        tailer.finish();
-                        break;
+        public void setPauser(Pauser pauser) {
+            this.pauser = pauser;
+        }
 
-                    } else {
-                        if (lastHeartbeat <= now) {
-                            sendSizeAndIndex(ChronicleTcp.IN_SYNC_LEN, ChronicleTcp.IDX_NONE);
+        public void setSubscriptionListener(SubscriptionListener subscriptionListener) {
+            this.subscriptionListener = subscriptionListener;
+        }
+
+        public void setMaxExcerptsPerMessage(int maxExcerptsPerMessage) {
+            this.maxExcerptsPerMessage = maxExcerptsPerMessage;
+        }
+
+        public void setHeartbeatIntervalMillis(long heartbeatIntervalMillis) {
+            this.heartbeatIntervalMillis = heartbeatIntervalMillis;
+        }
+
+        @Override
+        public void process(Bytes in, Bytes out, SessionDetailsProvider sessionDetailsProvider) {
+            processIncoming(in, out, sessionDetailsProvider);
+            processOutgoing(out, sessionDetailsProvider);
+        }
+
+        private void processIncoming(Bytes in, Bytes out, SessionDetailsProvider sessionDetailsProvider) {
+            while (in.remaining() >= 8) {
+
+                final long action = in.readLong();
+                try {
+                    switch ((int) action) {
+                        case (int) ChronicleTcp.ACTION_WITH_MAPPING:
+                            if (!onMapping(in, sessionDetailsProvider)) {
+                                unreadAction(in);
+                                return;
+                            }
                             break;
+                        case (int) ChronicleTcp.ACTION_SUBSCRIBE:
+                            if (!(subscribed = onSubscribe(in, out, sessionDetailsProvider))) {
+                                unreadAction(in); // wind back the long we read
+                                return;
+                            }
+                            break;
+                        case (int) ChronicleTcp.ACTION_UNSUBSCRIBE:
+                            if (!onUnsubscribe(in)) {
+                                unreadAction(in);
+                                return;
+                            }
+                            break;
+                        case (int) ChronicleTcp.ACTION_QUERY:
+                            if (!onQuery(in, out)) {
+                                in.position(in.position() - 8);
+                                return;
+                            }
+                            break;
+                        case (int) ChronicleTcp.ACTION_SUBMIT:
+                            if (!onSubmit(in, out, true)) {
+                                in.position(in.position() - 8);
+                                return;
+                            }
+                            break;
+                        case (int) ChronicleTcp.ACTION_SUBMIT_NOACK:
+                            if (!onSubmit(in, out, false)) {
+                                in.position(in.position() - 8);
+                                return;
+                            }
+                            break;
+                        default:
+                            throw new IOException("Unknown action received (" + action + ")");
+                    }
+                } catch (IOException e) {
+                    throw new TcpHandlingException(e);
+                }
+            }
+        }
+
+        private void unreadAction(Bytes in) {
+            in.position(in.position() - 8);
+        }
+
+        private void processOutgoing(Bytes out, SessionDetailsProvider sessionDetailsProvider) {
+            final long now = System.currentTimeMillis();
+
+            if (subscribed && !write(out, sessionDetailsProvider)) {
+                if (lastHeartbeat <= now && out.remaining() >= ChronicleTcp.HEADER_SIZE) {
+                    writeSizeAndIndex(out, ChronicleTcp.IN_SYNC_LEN, ChronicleTcp.IDX_NONE);
+                }
+            }
+        }
+
+        protected abstract boolean write(Bytes out, SessionDetailsProvider sessionDetailsProvider);
+
+        protected void writeSizeAndIndex(Bytes out, int size, long index) {
+            out.writeInt(size);
+            out.writeLong(index);
+            setLastHeartbeat();
+        }
+
+        protected void setLastHeartbeat() {
+            setLastHeartbeat(System.currentTimeMillis());
+        }
+
+        protected void setLastHeartbeat(long now) {
+            this.lastHeartbeat = now + heartbeatIntervalMillis;
+        }
+
+        @Override
+        public void onEndOfConnection() {
+
+        }
+
+        protected abstract boolean onSubmit(Bytes in, Bytes out, boolean ack) throws IOException;
+
+        protected abstract boolean onSubscribe(Bytes in, Bytes out, SessionDetailsProvider sessionDetailsProvider);
+
+        protected boolean onUnsubscribe(Bytes in) throws IOException {
+            if (in.remaining() >= 8) {
+                in.readLong(); // do nothing with the long
+                subscriptionListener.unSubscribed();
+                return true;
+            }
+            return false;
+        }
+
+        protected boolean onQuery(Bytes in, Bytes out) throws IOException {
+            if (in.remaining() >= 8) {
+                long index = in.readLong();
+                if (tailer.index(index)) {
+                    final long now = System.currentTimeMillis();
+                    setLastHeartbeat(now);
+
+                    while (true) {
+                        if (tailer.nextIndex()) {
+                            writeSizeAndIndex(out, ChronicleTcp.SYNC_IDX_LEN, tailer.index());
+                            tailer.finish();
+                            break;
+
+                        } else {
+                            if (lastHeartbeat <= now) {
+                                writeSizeAndIndex(out, ChronicleTcp.IN_SYNC_LEN, ChronicleTcp.IDX_NONE);
+                                break;
+                            }
                         }
                     }
+                } else {
+                    writeSizeAndIndex(out, ChronicleTcp.IN_SYNC_LEN, 0L);
                 }
-            } else {
-                sendSizeAndIndex(ChronicleTcp.IN_SYNC_LEN, 0L);
+                return true;
             }
 
-            return true;
+            return false;
         }
 
-        protected boolean onUnsubscribe(final SelectionKey key, long data) {
-            key.interestOps(key.interestOps() & ~SelectionKey.OP_WRITE);
-            return true;
+        protected boolean onMapping(Bytes in, SessionDetailsProvider sessionDetailsProvider) throws IOException {
+            if (in.remaining() >= 4) {
+                int size = in.readInt();
+                if (in.remaining() >= size) {
+                    MappingProvider mappingProvider = sessionDetailsProvider.get(MappingProvider.class);
+                    if (mappingProvider != null) {
+                        MappingFunction mappingFunction = in.readObject(MappingFunction.class);
+                        mappingProvider.withMapping(mappingFunction);
+                    }
+
+                    return true;
+                } else {
+                    in.position(in.position() - 4);
+                }
+            }
+            return false;
         }
-
-        protected abstract boolean onSubscribe(final SelectionKey key, long data) throws IOException;
-
-        protected abstract boolean onSubmit(final SelectionKey key, long size, boolean ack) throws IOException;
-
-        protected abstract boolean write(Object attachment) throws IOException;
 
         /**
          * applies a mapping if the mapping is not set to {@code}null{code}
          *
-         * @param source   the tailer for the mapping to be applied to
-         * @param attached the key attachment
+         * @param source          the tailer for the mapping to be applied to
+         * @param mappingProvider the key attachment
          * @return returns the tailer or the mapped bytes
          * @see
          */
         protected Bytes applyMapping(@NotNull final ExcerptTailer source,
-                                     @Nullable Object attached) {
-            if (attached == null) {
+                                     @Nullable MappingProvider mappingProvider) {
+            if (mappingProvider == null) {
                 return source;
             }
 
-            final MappingProvider mappingProvider = (MappingProvider) attached;
             final MappingFunction mappingFunction = mappingProvider.withMapping();
             if (mappingFunction == null) {
                 return source;
@@ -505,272 +587,301 @@ public abstract class SourceTcp {
 
             return withMappedBuffer.flip();
         }
-    }
 
-    // *************************************************************************
-    // SessionHandler - implementations
-    // *************************************************************************
+        protected void writePartial(Bytes out) {
+            int bytesToWrite = (int) Math.min(content.remaining(), out.remaining());
+            long contentLimit = content.limit();
+            out.write(content.limit(content.position() + bytesToWrite));
+            content.limit(contentLimit);
 
-    /**
-     * IndexedChronicle session handler
-     */
-    private class IndexedSessionHandler extends SessionHandler {
-        private long index;
+            state = content.remaining() > 0 ? TcpHandlerState.EXCERPT_INCOMPLETE : TcpHandlerState.EXCERPT_COMPLETE;
 
-        private IndexedSessionHandler(final @NotNull SocketChannel socketChannel) throws IOException {
-            super(socketChannel, SourceTcpHandler.indexed());
-            this.index = -1;
+            System.out.println("SourceTcp.writePartial; index: " + tailer.index() + ", state: " + state.name());
         }
 
-        @Override
-        protected boolean onSubscribe(final SelectionKey key, long data) throws IOException {
-            this.index = data;
-            if (this.index == ChronicleTcp.IDX_TO_START) {
-                this.index = -1;
-
-            } else if (this.index == ChronicleTcp.IDX_TO_END) {
-                this.index = tailer.toEnd().index();
-            }
-
-            sendSizeAndIndex(ChronicleTcp.SYNC_IDX_LEN, this.index);
-
-            key.interestOps(SelectionKey.OP_READ | SelectionKey.OP_WRITE);
-            return true;
+        protected boolean hasRoomForExcerpt(Bytes bytes, Bytes tailer) {
+            return hasRoomFor(bytes, tailer.remaining() + ChronicleTcp.HEADER_SIZE);
         }
 
-        @Override
-        protected boolean onSubmit(final SelectionKey key, long size, boolean ack) throws IOException {
-            if (ack) {
-                sendSizeAndIndex(ChronicleTcp.NACK_LEN, ChronicleTcp.IDX_NOT_SUPPORTED);
-            }
-
-            return true;
+        protected boolean hasRoomFor(Bytes bytes, long size) {
+            return bytes.remaining() >= size;
         }
 
-        @Override
-        protected boolean write(Object attached) throws IOException {
-            if (!tailer.index(index)) {
-                if (tailer.wasPadding()) {
-                    if (index >= 0) {
-                        sendSizeAndIndex(ChronicleTcp.PADDED_LEN, tailer.index());
+        public void setAppender(ExcerptAppender appender) {
+            this.appender = appender;
+        }
+
+        public static VanillaSourceTcpHandler vanilla() {
+            return new VanillaSourceTcpHandler();
+        }
+
+        public static class VanillaSourceTcpHandler extends SourceTcpHandler {
+
+            private boolean nextIndex;
+
+            @Override
+            protected boolean write(Bytes out, SessionDetailsProvider sessionDetailsProvider) {
+                if (state == TcpHandlerState.EXCERPT_INCOMPLETE) {
+                    writePartial(out);
+                    if (state == TcpHandlerState.EXCERPT_INCOMPLETE || out.remaining() <= ChronicleTcp.HEADER_SIZE) {
+                        // still incomplete but we have successfully written, no need to do anything else
+                        // or we're complete but unable to write anything else
+                        return true;
                     }
-
-                    index++;
                 }
 
-                pause();
 
-                if (running.get() && !tailer.index(index)) {
+                if (nextIndex) {
+                    if (!tailer.nextIndex()) {
+                        pauser.pause();
+                        if (!tailer.nextIndex()) {
+                            return false;
+                        }
+                    }
+                } else {
+                    if (!tailer.index(this.index)) {
+                        return false;
+
+                    } else {
+                        this.nextIndex = true;
+                    }
+                }
+
+                final MappingProvider mappingProvider = sessionDetailsProvider.get(MappingProvider.class);
+                content = applyMapping(tailer, mappingProvider);
+                int size = (int) content.remaining();
+
+                if (out.remaining() >= ChronicleTcp.HEADER_SIZE) {
+                    System.out.println("SourceTcp.write; index: " + tailer.index());
+                    writeSizeAndIndex(out, (int) content.limit(), tailer.index());
+
+                    // for large objects send one at a time.
+                    if (size > out.remaining()) {
+                        writePartial(out);
+                    } else {
+                        out.write(content);
+
+                        long previousIndex = tailer.index();
+                        long currentIndex;
+                        for (int count = maxExcerptsPerMessage; (count > 0) && tailer.nextIndex(); ) {
+                            currentIndex = tailer.index();
+                            content = applyMapping(tailer, mappingProvider);
+
+                            // if there is free space, copy another one.
+                            if (hasRoomForExcerpt(out, content)) {
+                                size = (int) content.limit();
+                                previousIndex = currentIndex;
+                                writeSizeAndIndex(out, size, currentIndex);
+                                out.write(content);
+                                count--;
+
+                                tailer.finish();
+                            } else {
+                                tailer.finish();
+                                // if there is no space, go back to the previous index
+                                tailer.index(previousIndex);
+                                break;
+                            }
+                        }
+
+                        state = TcpHandlerState.EXCERPT_COMPLETE;
+
+                    }
+
+                    return true;
+                } else {
+                    if (tailer.index(tailer.index())) {
+                        index = tailer.index();
+                        nextIndex = false;
+                    }
                     return false;
                 }
             }
 
-            pauseReset();
+            @Override
+            protected boolean onSubscribe(Bytes in, Bytes out, SessionDetailsProvider sessionDetailsProvider) {
+                if (in.remaining() >= 8) {
+                    this.index = in.readLong();
+                    if (this.index == ChronicleTcp.IDX_TO_START) {
+                        this.nextIndex = true;
+                        this.tailer = tailer.toStart();
+                        this.index = -1;
 
-            Bytes bytes = applyMapping(tailer, attached);
-            int size = (int) bytes.limit();
+                    } else if (this.index == ChronicleTcp.IDX_TO_END) {
+                        this.nextIndex = false;
+                        this.tailer = tailer.toEnd();
+                        this.index = tailer.index();
 
-            bytesOut.clear();
-            bytesOut.writeInt(size);
-            bytesOut.writeLong(tailer.index());
-
-            // for large objects send one at a time.
-            if (size > bytesOut.capacity() / 2) {
-                while (size > 0) {
-                    int minSize = (int) Math.min(size, bytesOut.remaining());
-                    bytesOut.write(bytes);
-                    // TODO
-//                    connection.write(bytesOut.flip());
-
-                    size -= minSize;
-                    if (size > 0) {
-                        bytesOut.clear();
-                    }
-                }
-            } else {
-                bytesOut.write(bytes);
-                for (int count = builder.maxExcerptsPerMessage(); (count > 0) && tailer.index(index + 1); ) {
-                    if (!tailer.wasPadding()) {
-                        bytes = applyMapping(tailer, attached);
-                        // if there is free space, copy another one.
-                        if (hasRoomForExcerpt(bytesOut, bytes)) {
-                            size = (int) bytes.limit();
-                            bytesOut.writeInt(size);
-                            bytesOut.writeLong(tailer.index());
-                            bytesOut.write(bytes);
-
-                            index++;
-                            count--;
-
-                            tailer.finish();
-
-                        } else {
-                            break;
+                        if (this.index == -1) {
+                            this.nextIndex = true;
+                            this.tailer = tailer.toStart();
+                            this.index = -1;
                         }
                     } else {
-                        if (hasRoomFor(bytesOut, ChronicleTcp.HEADER_SIZE)) {
-                            bytesOut.writeInt(ChronicleTcp.PADDED_LEN);
-                            bytesOut.writeLong(index);
+                        this.nextIndex = false;
+                    }
 
-                        } else {
-                            break;
+                    writeSizeAndIndex(out, ChronicleTcp.SYNC_IDX_LEN, this.index);
+
+                    subscriptionListener.subscribed();
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+
+            @Override
+            protected boolean onSubmit(Bytes in, Bytes out, boolean ack) throws IOException {
+                if (in.remaining() >= 8) {
+                    long size = in.readLong();
+                    if (in.remaining() >= size) {
+                        long inLimit = in.limit();
+                        in.limit(in.position() + size);
+
+                        try {
+                            appender.startExcerpt((int) size);
+                            appender.write(in);
+                            appender.finish();
+                        } finally {
+                            in.limit(inLimit);
+                        }
+
+                        pauser.unpause();
+
+                        if (ack) {
+                            writeSizeAndIndex(out, ChronicleTcp.ACK_LEN, appender.lastWrittenIndex());
+                        }
+
+                        return true;
+                    } else {
+                        in.position(in.position() - 8); // unread the size
+                    }
+                }
+                return false;
+            }
+
+        }
+
+        public static IndexedSourceTcpHandler indexed() {
+            return new IndexedSourceTcpHandler();
+        }
+
+        public static class IndexedSourceTcpHandler extends SourceTcpHandler {
+
+            @Override
+            protected boolean write(Bytes out, SessionDetailsProvider sessionDetailsProvider) {
+                if (state == TcpHandlerState.EXCERPT_INCOMPLETE) {
+                    writePartial(out);
+                    if (state == TcpHandlerState.EXCERPT_INCOMPLETE || out.remaining() <= ChronicleTcp.HEADER_SIZE) {
+                        // still incomplete but we have successfully written, no need to do anything else
+                        // or we're complete but unable to write anything else
+                        return true;
+                    }
+                }
+
+                if (!tailer.index(index)) {
+                    if (tailer.wasPadding()) {
+                        if (index >= 0) {
+                            writeSizeAndIndex(out, ChronicleTcp.PADDED_LEN, tailer.index());
                         }
 
                         index++;
                     }
-                }
 
-                // TODO
-//                connection.write(bytesOut.flip());
-            }
+                    pauser.pause();
 
-            if (bytesOut.remaining() > 0) {
-                throw new EOFException("Failed to send index=" + index);
-            }
-
-            index++;
-            return true;
-        }
-    }
-
-    /**
-     * VanillaChronicle session handler
-     */
-    private class VanillaSessionHandler extends SessionHandler {
-        private boolean nextIndex;
-        private long index;
-
-        private VanillaSessionHandler(final @NotNull SocketChannel socketChannel) throws IOException {
-            super(socketChannel, SourceTcpHandler.vanilla());
-
-            this.nextIndex = true;
-            this.index = -1;
-        }
-
-        @Override
-        protected boolean onSubscribe(final SelectionKey key, long data) throws IOException {
-            this.index = data;
-            if (this.index == ChronicleTcp.IDX_TO_START) {
-                this.nextIndex = true;
-                this.tailer = tailer.toStart();
-                this.index = -1;
-
-            } else if (this.index == ChronicleTcp.IDX_TO_END) {
-                this.nextIndex = false;
-                this.tailer = tailer.toEnd();
-                this.index = tailer.index();
-
-                if (this.index == -1) {
-                    this.nextIndex = true;
-                    this.tailer = tailer.toStart();
-                    this.index = -1;
-                }
-            } else {
-                this.nextIndex = false;
-            }
-
-            sendSizeAndIndex(ChronicleTcp.SYNC_IDX_LEN, this.index);
-
-            key.interestOps(SelectionKey.OP_READ | SelectionKey.OP_WRITE);
-            return false;
-        }
-
-        @Override
-        protected boolean onSubmit(final SelectionKey key, long size, boolean ack) throws IOException {
-            readUpTo((int) size);
-
-            appender.startExcerpt((int) size);
-            appender.write(readBuffer);
-            appender.finish();
-
-            pauser.unpause();
-
-            if (ack) {
-                sendSizeAndIndex(ChronicleTcp.ACK_LEN, appender.lastWrittenIndex());
-            }
-
-            return true;
-        }
-
-        @Override
-        protected boolean write(Object attached) throws IOException {
-            if (nextIndex) {
-                if (!tailer.nextIndex()) {
-                    pause();
-                    if (running.get() && !tailer.nextIndex()) {
+                    if (!tailer.index(index)) {
                         return false;
                     }
                 }
-            } else {
-                if (!tailer.index(this.index)) {
-                    return false;
 
-                } else {
-                    this.nextIndex = true;
-                }
-            }
+                pauser.reset();
 
-            pauseReset();
+                final MappingProvider mappingProvider = sessionDetailsProvider.get(MappingProvider.class);
+                content = applyMapping(tailer, mappingProvider);
+                int size = (int) content.limit();
 
-            Bytes bytes = applyMapping(tailer, attached);
-            int size = (int) bytes.limit();
+                if (out.remaining() >= ChronicleTcp.HEADER_SIZE) {
+                    writeSizeAndIndex(out, size, tailer.index());
 
-            bytesOut.clear();
-            bytesOut.writeInt(size);
-            bytesOut.writeLong(tailer.index());
-
-            // for large objects send one at a time.
-            if (size > bytesOut.limit() / 2) {
-                while (size > 0) {
-                    int minSize = (int) Math.min(size, bytesOut.remaining());
-                    bytesOut.write(bytes);
-                    // TODO
-//                    connection.write(bytesOut.flip());
-                    bytesOut.clear();
-
-                    size -= minSize;
-                    if (size > 0) {
-                        bytesOut.clear();
-                    }
-                }
-            } else {
-                bytesOut.write(bytes);
-
-                long previousIndex = tailer.index();
-                long currentIndex;
-                for (int count = builder.maxExcerptsPerMessage(); (count > 0) && tailer.nextIndex(); ) {
-                    currentIndex = tailer.index();
-                    bytes = applyMapping(tailer, attached);
-
-                    // if there is free space, copy another one.
-                    if (hasRoomForExcerpt(bytesOut, bytes)) {
-                        size = (int) bytes.limit();
-                        previousIndex = currentIndex;
-                        bytesOut.writeInt(size);
-                        bytesOut.writeLong(currentIndex);
-                        bytesOut.write(bytes);
-                        count--;
-
-                        tailer.finish();
-
+                    // for large objects send one at a time.
+                    if (content.remaining() > out.remaining()) {
+                        writePartial(out);
                     } else {
-                        tailer.finish();
-                        // if there is no space, go back to the previous index
-                        tailer.index(previousIndex);
-                        System.out.println("Winding back to previous index: " + previousIndex);
-                        break;
+                        out.write(content);
+                        for (int count = maxExcerptsPerMessage; (count > 0) && tailer.index(index + 1); ) {
+                            if (!tailer.wasPadding()) {
+                                content = applyMapping(tailer, mappingProvider);
+                                // if there is free space, copy another one.
+                                if (hasRoomForExcerpt(out, content)) {
+                                    size = (int) content.limit();
+                                    writeSizeAndIndex(out, size, tailer.index());
+                                    out.write(content);
+
+                                    index++;
+                                    count--;
+
+                                    tailer.finish();
+
+                                } else {
+                                    break;
+                                }
+                            } else {
+                                if (hasRoomFor(out, ChronicleTcp.HEADER_SIZE)) {
+                                    writeSizeAndIndex(out, ChronicleTcp.PADDED_LEN, index);
+                                } else {
+                                    break;
+                                }
+
+                                index++;
+                            }
+                        }
                     }
+
+                    index++;
+                    return true;
+                } else {
+                    return false;
+                }
+            }
+
+            @Override
+            protected boolean onSubscribe(Bytes in, Bytes out, SessionDetailsProvider sessionDetailsProvider) {
+                this.index = in.readLong();
+                if (this.index == ChronicleTcp.IDX_TO_START) {
+                    this.index = -1;
+
+                } else if (this.index == ChronicleTcp.IDX_TO_END) {
+                    this.index = tailer.toEnd().index();
                 }
 
-                // TODO
-//                connection.write(bytesOut.flip());
+                writeSizeAndIndex(out, ChronicleTcp.SYNC_IDX_LEN, this.index);
+
+                subscriptionListener.subscribed();
+                return true;
             }
 
-            if (bytesOut.remaining() > 0) {
-                throw new EOFException("Failed to send index=" + tailer.index());
-            }
+            @Override
+            protected boolean onSubmit(Bytes in, Bytes out, boolean ack) throws IOException {
+                if (ack) {
+                    writeSizeAndIndex(out, ChronicleTcp.NACK_LEN, ChronicleTcp.IDX_NOT_SUPPORTED);
+                }
 
-            return true;
+                return true;
+            }
         }
+
+        public interface SubscriptionListener {
+            void subscribed();
+
+            void unSubscribed();
+        }
+
     }
+
+    public enum TcpHandlerState {
+        EXCERPT_COMPLETE,
+        EXCERPT_INCOMPLETE
+    }
+
 }
