+++ chronicle/src/test/java/net/openhft/chronicle/tcp/ChronicleTcpTestBase.java
@@ -221,6 +221,7 @@ public class ChronicleTcpTestBase {
                         appender.writeLong(i);
                         appender.writeLong(i + 1);
                         appender.finish();
+                        System.out.println("Appender; i: " + i + ", index: " + appender.lastWrittenIndex() + ", " + i + ", " + (i + 1));
 
                         Thread.sleep(random.nextInt(10));
                     }
@@ -257,8 +258,11 @@ public class ChronicleTcpTestBase {
                         assertTrue("Timeout exceeded " + (end - start), (end - start) < timeout);
 
                         if(hasNext) {
-                            assertEquals(i, tailer.readLong());
-                            assertEquals(i + 1, tailer.readLong());
+                            long first = tailer.readLong();
+                            long second = tailer.readLong();
+                            System.out.println("Tailer; i: " + i + ", index: " + tailer.index() + ", " + first + ", " + second);
+                            assertEquals(i, first);
+                            assertEquals(i + 1, second);
                             i++;
 
                             tailer.finish();
